requirements.txt
# Core extraction
openpyxl>=3.0.0
python-pptx>=0.6.21

# Azure services
azure-storage-blob>=12.0.0
azure-identity>=1.0.0
azure-ai-documentintelligence>=1.0.0

# AI/LLM
openai>=1.0.0

# Image/PDF processing
pymupdf>=1.24.0

# Windows PowerPoint COM automation (optional)
comtypes>=1.4.0; sys_platform == 'win32'

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0
